feat: ApiKey lib
