+++ mysk-lib/Cargo.toml
@@ -11,6 +11,7 @@ env_logger = "0.11.0"
 serde_qs = "0.12.0"
 utoipa = { version = "4", features = ["actix_extras"] }
 # openssl={ version = "0.10", features = ["v110"] }
+
 sqlx = { version = "0.7.2", default-features = false, features = [
     "postgres",
     "uuid",
@@ -28,3 +29,6 @@ jsonwebtoken = "9.2.0"
 async-trait = "0.1.68"
 reqwest = { version = "0.11.23", features = ["json"] }
 futures = "0.3.17"
+bs58 = "0.5.0"
+sha2 = "0.10.8"
+rand = "0.8.4"
